Move AccessGrantError in its own module
